requirements.txt
mkdocs>=1.1
Pygments>=2.4
markdown>=3.2
pymdown-extensions>=7.0
mkdocs-material-extensions>=1.0
numpy
simpy
matplotlib
json5>=0.9.5
PyQt5==5.15.4
datetime
ARC-Alkali-Rydberg-Calculator>=3.0.5
legume-gme>=0.1.6